Add powerup feature to game
This commit adds a simple powerup feature for the player. It will show
up when the player gets a certain overall score. When a player has the
powerup, the ship will move faster and it will shoot two missles instead
of one. If the player collides with an alien, it will lose the powerup.